Fix `test_missing_options()` test removing too many options from the config file.
Since the `cfg` object was created *outside* the `missing_option` function, every time the function was called one *more* option was removed from the config file. This caused the test to always pass if the *first* option removed raised an error, even if other options were not raising the same error.