Add support for import/export operations
This closes #161

Export operation allows to also export ssh keys, although since they contain
sensible information, confirmation is asked to the user.

During the import operation, all attributes that need to be changed are
updated. Also, if ssh keys are imported, their paths in the Cluster and Node
classes is updated accordingly.