Refactor config parser code.
The new configuration file processor is built around a pipeline of
processing functions, which should make it simpler to extend, or
handle format changes and key or content renames in the config files.

The new (internal) API has a single entry point, namely factory
function `make_creator` which should be called with a list of
configuration fies to get a `Creator` object from whence to load or
create new instances of a cluster object.  (Further refactoring is
planned, to avoid the need of an intermediate `Creator` object, and
just move from config files to a Python data structure that can be
passed to `Cluster` constructors unchanged.)

Other notable changes implemented in the refactoring:

* Use `schematics` instead of `voluptuous` to perform configuration
  validation.

* While `%(...)s` are still supported through Python's `ConfigParser`,
  `${...}`-style substitutions are now available and preferred.  The
  `%`-style form (coming from Python's `ConfigParser`) is deprecated
  and will be possibly removed in ElastiCluster 2.0, as the same
  functionality will _not_ be available in JSON- or YAML-format
  configs.

* Environmental variables are expanded in *all* configuration values,
  via the `${...}`-style substitutions.

* `elasticluster_playbooks` is now a (more readable) alias for
  `ansible_pb_dir`; both available as `${...}` or `%`-style substitutions.

* The `playbook_path` configuration optiions defaults to ElastiCluster's `site.yml`

* The `image_userdata` configuration option is now supported in both
  node definition and `[login/*]` sections (the former takes precedence).

* Configuration options `price` and `timeout` (in the EC2-type
  `[cloud/*]` sections),  and `zone` and `network` (in GCE `[cloud/*]`
  sections) are read and provide defaults as documented.