EC2: Use AWS credentials as written in the config file.
This should solve a few recent bug reports, where it was noted that
the AWS access credentials had to be defined as environmental
variable (`AWS_ACCESS_KEY_ID` and `AWS_SECRET_KEY`) in addition to
being in the config file: the reason being that ElastiCluster code was
passing the AWS credentials to `boto.ec2.get_region()` which basically
silently ignores them, instead of using `boto.ec2.connect_to_region()`
which does the right thing.

Many thanks to Sergio Mafra and Hatef Monajemi for reporting the issue!