Try to fix concurrency error in final status reporting.
It seems that each Ansible fork operates concurrently on the
`done.log` file and the `lineinfile:` is not atomic; so for
sufficiently large clusters a few nodes are randomly skipped.

Switching to separate reports for each nodes should solve the problem,
since a POSIX filesystem will not have issues with concurrent creation
of files by independent processes.