Use workflow dispatch to trigger action workflows (#692)
Try two-stage dispatch for solving the dependency issue in GH workflows:

- there's one "receiver" workflow that runs upon GH events like push, release, etc.
- the event workflow triggers other workflows which perform actual actions: e.g., run unit tests, build Docker image, etc.

This solves the dependency and code sharing problem with workflows: there's a single definition of Unit Tests that is shared by all events that need it, *and* we can sequence operations (like build Docker image only if Unit Tests were successful).

The downside is that there's a lot of boilerplate code. (Turn the trigger code into an action? but there's no such thing as repo-local actions, so a separate repository would be needed just for ~20 lines of shell script.)

Note: the standard GITHUB_TOKEN secret cannot launch other workflows; this is explained in:
<https://docs.github.com/en/free-pro-team@latest/actions/reference/events-that-trigger-workflows#triggering-new-workflows-using-a-personal-access-token>